move logic out of routes and into handlers and services
