+++ internal/features/index/routes.go
@@ -1,374 +1,39 @@
 package index
 
 import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"strconv"
-	"time"
-
-	"github.com/delaneyj/toolbelt"
 	"github.com/delaneyj/toolbelt/embeddednats"
-	"github.com/starfederation/datastar-go/datastar"
-
 	"github.com/go-chi/chi/v5"
 	"github.com/gorilla/sessions"
-	"github.com/nats-io/nats.go/jetstream"
-	"github.com/samber/lo"
-	"github.com/zangster300/northstar/internal/features/common/components"
-	"github.com/zangster300/northstar/internal/features/index/pages"
+	"github.com/zangster300/northstar/internal/features/index/services"
 )
 
 func SetupRoutes(router chi.Router, store sessions.Store, ns *embeddednats.Server) error {
-	nc, err := ns.Client()
+	todoService, err := services.NewTodoService(ns, store)
 	if err != nil {
-		return fmt.Errorf("error creating nats client: %w", err)
+		return err
 	}
 
-	js, err := jetstream.New(nc)
-	if err != nil {
-		return fmt.Errorf("error creating jetstream client: %w", err)
-	}
+	handlers := NewHandlers(todoService)
 
-	kv, err := js.CreateOrUpdateKeyValue(context.Background(), jetstream.KeyValueConfig{
-		Bucket:      "todos",
-		Description: "Datastar Todos",
-		Compression: true,
-		TTL:         time.Hour,
-		MaxBytes:    16 * 1024 * 1024,
-	})
-
-	if err != nil {
-		return fmt.Errorf("error creating key value: %w", err)
-	}
-
-	saveMVC := func(ctx context.Context, sessionID string, mvc *components.TodoMVC) error {
-		b, err := json.Marshal(mvc)
-		if err != nil {
-			return fmt.Errorf("failed to marshal mvc: %w", err)
-		}
-		if _, err := kv.Put(ctx, sessionID, b); err != nil {
-			return fmt.Errorf("failed to put key value: %w", err)
-		}
-		return nil
-	}
-
-	resetMVC := func(mvc *components.TodoMVC) {
-		mvc.Mode = components.TodoViewModeAll
-		mvc.Todos = []*components.Todo{
-			{Text: "Learn any backend language", Completed: true},
-			{Text: "Learn Datastar", Completed: false},
-			{Text: "Create Hypermedia", Completed: false},
-			{Text: "???", Completed: false},
-			{Text: "Profit", Completed: false},
-		}
-		mvc.EditingIdx = -1
-	}
-
-	mvcSession := func(w http.ResponseWriter, r *http.Request) (string, *components.TodoMVC, error) {
-		ctx := r.Context()
-		sessionID, err := upsertSessionID(store, r, w)
-		if err != nil {
-			return "", nil, fmt.Errorf("failed to get session id: %w", err)
-		}
-
-		mvc := &components.TodoMVC{}
-		if entry, err := kv.Get(ctx, sessionID); err != nil {
-			if err != jetstream.ErrKeyNotFound {
-				return "", nil, fmt.Errorf("failed to get key value: %w", err)
-			}
-			resetMVC(mvc)
-
-			if err := saveMVC(ctx, sessionID, mvc); err != nil {
-				return "", nil, fmt.Errorf("failed to save mvc: %w", err)
-			}
-		} else {
-			if err := json.Unmarshal(entry.Value(), mvc); err != nil {
-				return "", nil, fmt.Errorf("failed to unmarshal mvc: %w", err)
-			}
-		}
-		return sessionID, mvc, nil
-	}
-
-	router.Get("/", func(w http.ResponseWriter, r *http.Request) {
-		if err := pages.IndexPage("Northstar").Render(r.Context(), w); err != nil {
-			http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-		}
-	})
+	router.Get("/", handlers.IndexPage)
 
 	router.Route("/api", func(apiRouter chi.Router) {
 		apiRouter.Route("/todos", func(todosRouter chi.Router) {
-			todosRouter.Get("/", func(w http.ResponseWriter, r *http.Request) {
-
-				sessionID, mvc, err := mvcSession(w, r)
-				if err != nil {
-					http.Error(w, err.Error(), http.StatusInternalServerError)
-					return
-				}
-
-				sse := datastar.NewSSE(w, r)
-
-				// Watch for updates
-				ctx := r.Context()
-				watcher, err := kv.Watch(ctx, sessionID)
-				if err != nil {
-					http.Error(w, err.Error(), http.StatusInternalServerError)
-					return
-				}
-				defer watcher.Stop()
-
-				for {
-					select {
-					case <-ctx.Done():
-						return
-					case entry := <-watcher.Updates():
-						if entry == nil {
-							continue
-						}
-						if err := json.Unmarshal(entry.Value(), mvc); err != nil {
-							http.Error(w, err.Error(), http.StatusInternalServerError)
-							return
-						}
-						c := components.TodosMVCView(mvc)
-						if err := sse.PatchElementTempl(c); err != nil {
-							if err := sse.ConsoleError(err); err != nil {
-								http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-							}
-							return
-						}
-					}
-				}
-			})
-
-			todosRouter.Put("/reset", func(w http.ResponseWriter, r *http.Request) {
-				sessionID, mvc, err := mvcSession(w, r)
-				if err != nil {
-					http.Error(w, err.Error(), http.StatusInternalServerError)
-					return
-				}
-
-				resetMVC(mvc)
-				if err := saveMVC(r.Context(), sessionID, mvc); err != nil {
-					http.Error(w, err.Error(), http.StatusInternalServerError)
-					return
-				}
-			})
-
-			todosRouter.Put("/cancel", func(w http.ResponseWriter, r *http.Request) {
-
-				sessionID, mvc, err := mvcSession(w, r)
-				sse := datastar.NewSSE(w, r)
-				if err != nil {
-					if err := sse.ConsoleError(err); err != nil {
-						http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-					}
-					return
-				}
-
-				mvc.EditingIdx = -1
-				if err := saveMVC(r.Context(), sessionID, mvc); err != nil {
-					if err := sse.ConsoleError(err); err != nil {
-						http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-					}
-					return
-				}
-			})
-
-			todosRouter.Put("/mode/{mode}", func(w http.ResponseWriter, r *http.Request) {
-
-				sessionID, mvc, err := mvcSession(w, r)
-				if err != nil {
-					http.Error(w, err.Error(), http.StatusInternalServerError)
-					return
-				}
-
-				modeStr := chi.URLParam(r, "mode")
-				modeRaw, err := strconv.Atoi(modeStr)
-				if err != nil {
-					http.Error(w, err.Error(), http.StatusBadRequest)
-					return
-				}
-
-				mode := components.TodoViewMode(modeRaw)
-				if mode < components.TodoViewModeAll || mode > components.TodoViewModeCompleted {
-					http.Error(w, "invalid mode", http.StatusBadRequest)
-					return
-				}
-
-				mvc.Mode = mode
-				if err := saveMVC(r.Context(), sessionID, mvc); err != nil {
-					http.Error(w, err.Error(), http.StatusInternalServerError)
-					return
-				}
-			})
+			todosRouter.Get("/", handlers.TodosSSE)
+			todosRouter.Put("/reset", handlers.ResetTodos)
+			todosRouter.Put("/cancel", handlers.CancelEdit)
+			todosRouter.Put("/mode/{mode}", handlers.SetMode)
 
 			todosRouter.Route("/{idx}", func(todoRouter chi.Router) {
-				routeIndex := func(w http.ResponseWriter, r *http.Request) (int, error) {
-					idx := chi.URLParam(r, "idx")
-					i, err := strconv.Atoi(idx)
-					if err != nil {
-						http.Error(w, err.Error(), http.StatusBadRequest)
-						return 0, err
-					}
-					return i, nil
-				}
-
-				todoRouter.Post("/toggle", func(w http.ResponseWriter, r *http.Request) {
-					sessionID, mvc, err := mvcSession(w, r)
-
-					sse := datastar.NewSSE(w, r)
-					if err != nil {
-						if err := sse.ConsoleError(err); err != nil {
-							http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-						}
-						return
-					}
-
-					i, err := routeIndex(w, r)
-					if err != nil {
-						if err := sse.ConsoleError(err); err != nil {
-							http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-						}
-						return
-					}
-
-					if i < 0 {
-						setCompletedTo := false
-						for _, todo := range mvc.Todos {
-							if !todo.Completed {
-								setCompletedTo = true
-								break
-							}
-						}
-						for _, todo := range mvc.Todos {
-							todo.Completed = setCompletedTo
-						}
-					} else {
-						todo := mvc.Todos[i]
-						todo.Completed = !todo.Completed
-					}
-
-					if err := saveMVC(r.Context(), sessionID, mvc); err != nil {
-						http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-					}
-				})
-
+				todoRouter.Post("/toggle", handlers.ToggleTodo)
 				todoRouter.Route("/edit", func(editRouter chi.Router) {
-					editRouter.Get("/", func(w http.ResponseWriter, r *http.Request) {
-						sessionID, mvc, err := mvcSession(w, r)
-						if err != nil {
-							http.Error(w, err.Error(), http.StatusInternalServerError)
-							return
-						}
-
-						i, err := routeIndex(w, r)
-						if err != nil {
-							return
-						}
-
-						mvc.EditingIdx = i
-						if err := saveMVC(r.Context(), sessionID, mvc); err != nil {
-							http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-						}
-					})
-
-					editRouter.Put("/", func(w http.ResponseWriter, r *http.Request) {
-						type Store struct {
-							Input string `json:"input"`
-						}
-						store := &Store{}
-
-						if err := datastar.ReadSignals(r, store); err != nil {
-							http.Error(w, err.Error(), http.StatusBadRequest)
-							return
-						}
-
-						if store.Input == "" {
-							return
-						}
-
-						sessionID, mvc, err := mvcSession(w, r)
-						if err != nil {
-							http.Error(w, err.Error(), http.StatusInternalServerError)
-							return
-						}
-
-						i, err := routeIndex(w, r)
-						if err != nil {
-							return
-						}
-
-						if i >= 0 {
-							mvc.Todos[i].Text = store.Input
-						} else {
-							mvc.Todos = append(mvc.Todos, &components.Todo{
-								Text:      store.Input,
-								Completed: false,
-							})
-						}
-						mvc.EditingIdx = -1
-
-						if err := saveMVC(r.Context(), sessionID, mvc); err != nil {
-							http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-						}
-
-					})
+					editRouter.Get("/", handlers.StartEdit)
+					editRouter.Put("/", handlers.SaveEdit)
 				})
-
-				todoRouter.Delete("/", func(w http.ResponseWriter, r *http.Request) {
-					i, err := routeIndex(w, r)
-					if err != nil {
-						return
-					}
-
-					sessionID, mvc, err := mvcSession(w, r)
-					if err != nil {
-						http.Error(w, err.Error(), http.StatusInternalServerError)
-						return
-					}
-
-					if i >= 0 {
-						mvc.Todos = append(mvc.Todos[:i], mvc.Todos[i+1:]...)
-					} else {
-						mvc.Todos = lo.Filter(mvc.Todos, func(todo *components.Todo, i int) bool {
-							return !todo.Completed
-						})
-					}
-					if err := saveMVC(r.Context(), sessionID, mvc); err != nil {
-						http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-					}
-				})
+				todoRouter.Delete("/", handlers.DeleteTodo)
 			})
 		})
 	})
 
 	return nil
-}
-
-func MustJSONMarshal(v any) string {
-	b, err := json.MarshalIndent(v, "", " ")
-	if err != nil {
-		panic(err)
-	}
-	return string(b)
-}
-
-func upsertSessionID(store sessions.Store, r *http.Request, w http.ResponseWriter) (string, error) {
-	sess, err := store.Get(r, "connections")
-	if err != nil {
-		return "", fmt.Errorf("failed to get session: %w", err)
-	}
-
-	id, ok := sess.Values["id"].(string)
-
-	if !ok {
-		id = toolbelt.NextEncodedID()
-		sess.Values["id"] = id
-		if err := sess.Save(r, w); err != nil {
-			return "", fmt.Errorf("failed to save session: %w", err)
-		}
-	}
-
-	return id, nil
 }
